Remove $DFLAGS build type (but keep supporting the $DFLAGS variable)
While it is useful and expected that dub would honor the $DFLAGS
variable, it should not drop *all* flags that it adds on its own.

The most egregious way this is manifesting is that:
- `dub test` - runs unittests (duh)
- `DFLAGS= dub test` - does not run a single test

This is because dub would treat the presence (even if empty) of
$DFLAGS as flags to override the current buildType. This has been
changed and now $DFLAGS simply are appended to the flags derived from
the buildType. If users actually want to prevent dub from passing any
special flags they can do so with `dub test --build=plain`

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>