compilers/dmd: Determine the backend version for gdmd
This changes makes the compiler version for gdmd match the gdc version
instead of the dmd FE version that gdc had been using. This matches
what happens for ldmd2 which mirrors the ldc2 version.

Specifically, now the versions look like:
- ldmd2-1.40 :: 1.40.X
- ldc2-1.40 :: 1.40.X
- gdc-14 :: 14.X.Y
- gdmd-14 :: 14.X.Y

When, previously, gdmd-14 would result in a version like 2.108.X

Signed-off-by: Andrei Horodniceanu <a.horodniceanu@proton.me>